fixed, profile update bug, update imeditely now, to test loading for profile on IOS
